+++ libraries/aircraft-simulator/src/aircraft.cpp
@@ -61,7 +61,7 @@ void Aircraft::checkAcceleration()
     }
 }
 
-float Aircraft::sendSignal(Tower tower, const OurVector<3>& current_position)
+double Aircraft::sendSignal(Tower tower, const OurVector<3>& current_position)
 {
     return tower.calculateTime(current_position);
 }
